Remove old metadata stuff
