+++ crates/fuel-core/src/service/adapters/gas_price_adapters.rs
@@ -1,2 +1,1 @@
-pub mod fuel_gas_price_metadata;
 pub mod fuel_l2_block_source;
